add logout command

+++ linenoise.c
@@ -903,7 +903,7 @@ void linenoisePrintKeyCodes(void) {
         if (memcmp(quit,"quit",sizeof(quit)) == 0) break;
 
         printf("'%c' %02x (%d) (type quit to exit)\n",
-            isprint(c) ? c : '?', (int)c, (int)c);
+            isprint((int)c) ? c : '?', (int)c, (int)c);
         printf("\x1b[0G"); /* Go left edge manually, we are in raw mode. */
         fflush(stdout);
     }
